Add option to ignore job dependencies within command groups
* Command groups are particularly helpful to specifying jobs with
dependencies and for grouping output for reporting. However, in some
cases, enforcing job dependency is unnecessary and can increase build
times. To add flexibility to command groups, I've added a flag
('--ignore-job-dependencies') that can be used to change the behavior of
ciRun.py. When this flag is set to True, the slurmGcpWorkflow in ciRun.py
will not add the `--dependency` flag to the sbatch command so that all
jobs can run asynchronously. This can reduce time spent in testing,
providing someone's cluster and GCP project have sufficient quota.